+++ accounting_book_app/accounting_book_appApp.swift
@@ -43,7 +43,7 @@ struct accounting_book_appApp: App {
      }
     var body: some Scene {
         WindowGroup {
-            initPage()
+            FirstView()
             //ItemView(myGroupData: myGroup, myItemData: myItem, peoplePay: pPay)
         }
     }
